Split out sub classes of GameTransaction
