+++ src/main/java/org/spongepowered/common/event/tracking/context/transaction/GameTransaction.java
@@ -25,69 +25,28 @@
 package org.spongepowered.common.event.tracking.context.transaction;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.LinkedListMultimap;
-import com.google.common.collect.ListMultimap;
-import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.server.ServerWorld;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.framework.qual.DefaultQualifier;
-import org.spongepowered.api.Sponge;
-import org.spongepowered.api.block.BlockSnapshot;
-import org.spongepowered.api.data.Transaction;
 import org.spongepowered.api.event.Cancellable;
 import org.spongepowered.api.event.Cause;
 import org.spongepowered.api.event.CauseStackManager;
 import org.spongepowered.api.event.Event;
-import org.spongepowered.api.event.EventContext;
-import org.spongepowered.api.event.EventContextKey;
-import org.spongepowered.api.event.SpongeEventFactory;
-import org.spongepowered.api.event.block.ChangeBlockEvent;
-import org.spongepowered.api.event.block.NotifyNeighborBlockEvent;
-import org.spongepowered.api.util.Direction;
-import org.spongepowered.api.world.BlockChangeFlag;
-import org.spongepowered.api.world.BlockChangeFlags;
-import org.spongepowered.api.world.LocatableBlock;
-import org.spongepowered.common.block.SpongeBlockSnapshot;
-import org.spongepowered.common.block.SpongeBlockSnapshotBuilder;
-import org.spongepowered.common.bridge.tileentity.TileEntityBridge;
-import org.spongepowered.common.event.tracking.BlockChangeFlagManager;
-import org.spongepowered.common.event.tracking.IPhaseState;
 import org.spongepowered.common.event.tracking.PhaseContext;
-import org.spongepowered.common.event.tracking.PhaseTracker;
-import org.spongepowered.common.event.tracking.context.transaction.effect.BlockAddedEffect;
-import org.spongepowered.common.event.tracking.context.transaction.effect.CheckBlockPostPlacementIsSameEffect;
-import org.spongepowered.common.event.tracking.context.transaction.effect.ChunkChangeCompleteEffect;
-import org.spongepowered.common.event.tracking.context.transaction.effect.OldBlockOnReplaceEffect;
-import org.spongepowered.common.event.tracking.context.transaction.effect.RefreshOldTileEntityOnChunkChangeEffect;
-import org.spongepowered.common.event.tracking.context.transaction.effect.SetBlockToChunkSectionEffect;
-import org.spongepowered.common.event.tracking.context.transaction.effect.UpdateChunkLightManagerEffect;
-import org.spongepowered.common.event.tracking.context.transaction.effect.UpdateHeightMapEffect;
-import org.spongepowered.common.event.tracking.context.transaction.effect.UpdateOrCreateNewTileEntityPostPlacementEffect;
-import org.spongepowered.common.event.tracking.context.transaction.pipeline.ChunkPipeline;
-import org.spongepowered.common.util.Constants;
 import org.spongepowered.common.util.PrettyPrinter;
 import org.spongepowered.common.world.BlockChange;
-import org.spongepowered.common.world.SpongeBlockChangeFlag;
-import org.spongepowered.common.world.SpongeLocatableBlockBuilder;
-import org.spongepowered.math.vector.Vector3i;
 
-import java.util.Collections;
 import java.util.Deque;
-import java.util.EnumMap;
 import java.util.LinkedList;
-import java.util.List;
 import java.util.Optional;
 import java.util.StringJoiner;
 import java.util.function.BiConsumer;
-import java.util.function.Supplier;
 
 @DefaultQualifier(NonNull.class)
 public abstract class GameTransaction<E extends Event & Cancellable> {
-    private static final int EVENT_COUNT = BlockChange.values().length + 1;
 
     // State definitions
     final BlockPos affectedPosition;
@@ -158,523 +117,4 @@ public abstract class GameTransaction<E extends Event & Cancellable> {
 
     public abstract boolean markCancelledTransactions(E event, ImmutableList<? extends GameTransaction<E>> transactions);
 
-    static abstract class BlockEventBasedTransaction extends GameTransaction<ChangeBlockEvent> {
-
-        BlockEventBasedTransaction(final BlockPos affectedPosition, final BlockState originalState) {
-            super(affectedPosition, originalState);
-        }
-
-        @SuppressWarnings({"unchecked", "rawtypes"})
-        @Override
-        public final ChangeBlockEvent generateEvent(final PhaseContext<@NonNull ?> context,
-            final ImmutableList<GameTransaction<ChangeBlockEvent>> transactions,
-            final Cause currentCause
-        ) {
-            final ListMultimap<BlockPos, SpongeBlockSnapshot> positions = LinkedListMultimap.create();
-            for (final GameTransaction<@NonNull ?> transaction : transactions) {
-                if (!positions.containsKey(transaction.affectedPosition)) {
-                    positions.put(transaction.affectedPosition, ((BlockEventBasedTransaction) transaction).getOriginalSnapshot());
-                }
-                positions.put(transaction.affectedPosition, ((BlockEventBasedTransaction) transaction).getResultingSnapshot());
-            }
-
-            final ImmutableList<Transaction<BlockSnapshot>>[] transactionArrays = new ImmutableList[GameTransaction.EVENT_COUNT];
-            final ImmutableList.Builder<Transaction<BlockSnapshot>>[] transactionBuilders = new ImmutableList.Builder[GameTransaction.EVENT_COUNT];
-            for (int i = 0; i < GameTransaction.EVENT_COUNT; i++) {
-                transactionBuilders[i] = new ImmutableList.Builder<>();
-            }
-            // Bug is here- use the multimap
-            final ImmutableList<Transaction<BlockSnapshot>> eventTransactions = transactions.stream()
-                .map(transaction -> {
-                    final List<SpongeBlockSnapshot> snapshots = positions.get(transaction.affectedPosition);
-                    final SpongeBlockSnapshot original = snapshots.get(0);
-                    final SpongeBlockSnapshot result = snapshots.get(snapshots.size() - 1);
-                    final ImmutableList<BlockSnapshot> intermediary;
-                    if (snapshots.size() > 2) {
-                        intermediary = ImmutableList.copyOf(snapshots.subList(1, snapshots.size() - 2));
-                    } else {
-                        intermediary = ImmutableList.of();
-                    }
-                    final Transaction<BlockSnapshot> eventTransaction = new Transaction<>(original, result, intermediary);
-                    transactionBuilders[original.blockChange.ordinal()].add(eventTransaction);
-                    return eventTransaction;
-                }).collect(ImmutableList.toImmutableList());
-            for (int i = 0; i < GameTransaction.EVENT_COUNT; i++) {
-                transactionArrays[i] = transactionBuilders[i].build();
-            }
-            final @Nullable ChangeBlockEvent[] mainEvents = new ChangeBlockEvent[BlockChange.values().length];
-            for (final BlockChange blockChange : BlockChange.values()) {
-                if (blockChange == BlockChange.DECAY) { // Decay takes place after.
-                    continue;
-                }
-                if (!transactionArrays[blockChange.ordinal()].isEmpty()) {
-                    final ChangeBlockEvent event = blockChange.createEvent(currentCause, transactionArrays[blockChange.ordinal()]);
-                    mainEvents[blockChange.ordinal()] = event;
-                    Sponge.getEventManager().post(event);
-                }
-            }
-            if (!transactionArrays[BlockChange.DECAY.ordinal()].isEmpty()) { // Needs to be placed into iterateChangeBlockEvents
-                final ChangeBlockEvent event = BlockChange.DECAY.createEvent(currentCause, transactionArrays[BlockChange.DECAY.ordinal()]);
-                mainEvents[BlockChange.DECAY.ordinal()] = event;
-                Sponge.getEventManager().post(event);
-            }
-            final Cause causeToUse;
-            if (((IPhaseState) context.state).shouldProvideModifiers(context)) {
-                final Cause.Builder builder = Cause.builder().from(currentCause);
-                final EventContext.Builder modified = EventContext.builder();
-                modified.from(currentCause.getContext());
-                for (final BlockChange blockChange : BlockChange.values()) {
-                    final @Nullable ChangeBlockEvent mainEvent = mainEvents[blockChange.ordinal()];
-                    if (mainEvent != null) {
-                        builder.append(mainEvent);
-                        modified.add((EventContextKey<? super ChangeBlockEvent>) blockChange.getKey(), mainEvent);
-                    }
-                }
-                causeToUse = builder.build(modified.build());
-            } else {
-                causeToUse = currentCause;
-            }
-            return SpongeEventFactory.createChangeBlockEventPost(
-                causeToUse,
-                eventTransactions
-            );
-        }
-
-        protected abstract SpongeBlockSnapshot getResultingSnapshot();
-
-        protected abstract SpongeBlockSnapshot getOriginalSnapshot();
-
-        @Override
-        public boolean canBatchWith(final @Nullable GameTransaction<@NonNull ?> next) {
-            return next instanceof BlockEventBasedTransaction;
-        }
-
-        @Override
-        public final boolean markCancelledTransactions(final ChangeBlockEvent event,
-            final ImmutableList<? extends GameTransaction<ChangeBlockEvent>> blockTransactions
-        ) {
-            boolean cancelledAny = false;
-            for (final Transaction<BlockSnapshot> transaction: event.getTransactions()) {
-                if (!transaction.isValid()) {
-                    cancelledAny = true;
-                    for (final GameTransaction<ChangeBlockEvent> gameTransaction : blockTransactions) {
-                        final Vector3i position = transaction.getOriginal().getPosition();
-                        final BlockPos affectedPosition = gameTransaction.affectedPosition;
-                        if (position.getX() == affectedPosition.getX()
-                            && position.getY() == affectedPosition.getY()
-                            && position.getZ() == affectedPosition.getZ()
-                        ) {
-                            gameTransaction.markCancelled();
-                        }
-                    }
-                }
-            }
-
-            return cancelledAny;
-        }
-    }
-
-    @DefaultQualifier(NonNull.class)
-    public static final class AddTileEntity extends BlockEventBasedTransaction {
-
-        final TileEntity added;
-        final SpongeBlockSnapshot oldSnapshot;
-        final SpongeBlockSnapshot addedSnapshot;
-
-        AddTileEntity(final TileEntity added,
-            final SpongeBlockSnapshot attachedSnapshot,
-            final SpongeBlockSnapshot existing
-        ) {
-            super(existing.getBlockPos(), (BlockState) existing.getState());
-            this.added = added;
-            this.addedSnapshot = attachedSnapshot;
-            this.oldSnapshot = existing;
-        }
-
-        @Override
-        public Optional<BiConsumer<PhaseContext<@NonNull ?>, CauseStackManager.StackFrame>> getFrameMutator() {
-            return Optional.empty();
-        }
-
-        @Override
-        public void addToPrinter(final PrettyPrinter printer) {
-            printer.add("AddTileEntity")
-                .addWrapped(120, " %s : %s", this.affectedPosition, ((TileEntityBridge) this.added).bridge$getPrettyPrinterString());
-        }
-
-        @Override
-        public void restore() {
-            this.oldSnapshot.restore(true, BlockChangeFlags.NONE);
-        }
-
-        @Override
-        protected SpongeBlockSnapshot getResultingSnapshot() {
-            return this.addedSnapshot;
-        }
-
-        @Override
-        protected SpongeBlockSnapshot getOriginalSnapshot() {
-            return this.addedSnapshot;
-        }
-    }
-
-    @DefaultQualifier(NonNull.class)
-    public static final class RemoveTileEntity extends BlockEventBasedTransaction {
-
-        final TileEntity removed;
-        final SpongeBlockSnapshot tileSnapshot;
-
-        RemoveTileEntity(final TileEntity removed, final SpongeBlockSnapshot attachedSnapshot) {
-            super(attachedSnapshot.getBlockPos(), (BlockState) attachedSnapshot.getState());
-            this.removed = removed;
-            this.tileSnapshot = attachedSnapshot;
-        }
-
-        @Override
-        public Optional<BiConsumer<PhaseContext<@NonNull ?>, CauseStackManager.StackFrame>> getFrameMutator() {
-            return Optional.empty();
-        }
-
-        @Override
-        public void addToPrinter(final PrettyPrinter printer) {
-            printer.add("RemoveTileEntity")
-                .add(" %s : %s", this.affectedPosition, ((TileEntityBridge) this.removed).bridge$getPrettyPrinterString())
-                .add(" %s : %s", this.affectedPosition, this.originalState)
-            ;
-        }
-
-        @Override
-        public void restore() {
-            this.tileSnapshot.restore(true, BlockChangeFlags.NONE);
-        }
-
-        @Override
-        protected SpongeBlockSnapshot getResultingSnapshot() {
-            return SpongeBlockSnapshotBuilder.pooled()
-                .world((ServerWorld) this.removed.getWorld())
-                .position(new Vector3i(this.affectedPosition.getX(), this.affectedPosition.getY(), this.affectedPosition.getZ()))
-                .blockState(this.originalState)
-                .build()
-                ;
-        }
-
-        @Override
-        protected SpongeBlockSnapshot getOriginalSnapshot() {
-            return this.tileSnapshot;
-        }
-    }
-
-    @DefaultQualifier(NonNull.class)
-    public static final class ReplaceTileEntity extends BlockEventBasedTransaction {
-
-        final TileEntity added;
-        final TileEntity removed;
-        final SpongeBlockSnapshot removedSnapshot;
-
-        ReplaceTileEntity(final TileEntity added, final TileEntity removed,
-            final SpongeBlockSnapshot attachedSnapshot
-        ) {
-            super(attachedSnapshot.getBlockPos(), (BlockState) attachedSnapshot.getState());
-            this.added = added;
-            this.removed = removed;
-            this.removedSnapshot = attachedSnapshot;
-        }
-
-        @Override
-        public boolean acceptTileAddition(final TileEntity tileEntity) {
-            if (this.added == tileEntity) {
-                return true;
-            }
-            return super.acceptTileAddition(tileEntity);
-        }
-
-        @Override
-        public void restore() {
-            this.removedSnapshot.restore(true, BlockChangeFlags.NONE);
-        }
-
-        @Override
-        public Optional<BiConsumer<PhaseContext<@NonNull ?>, CauseStackManager.StackFrame>> getFrameMutator() {
-            return Optional.empty();
-        }
-
-        @Override
-        public void addToPrinter(final PrettyPrinter printer) {
-            printer.add("ReplaceTileEntity")
-                .add(" %s : %s", "Position", this.affectedPosition)
-                .add(" %s : %s", "Added", this.added)
-                .add(" %s : %s", "Removed", this.removed)
-            ;
-        }
-
-        @Override
-        protected SpongeBlockSnapshot getResultingSnapshot() {
-            return SpongeBlockSnapshotBuilder.pooled()
-                .from(this.removedSnapshot)
-                .tileEntity(this.added)
-                .build()
-                ;
-        }
-
-        @Override
-        protected SpongeBlockSnapshot getOriginalSnapshot() {
-            return this.removedSnapshot;
-        }
-    }
-
-    @DefaultQualifier(NonNull.class)
-    public static final class ChangeBlock extends BlockEventBasedTransaction {
-
-        final SpongeBlockSnapshot original;
-        final int originalOpacity;
-        final BlockState newState;
-        final SpongeBlockChangeFlag blockChangeFlag;
-        @Nullable public TileEntity queuedRemoval;
-        @Nullable public TileEntity queuedAdd;
-
-        ChangeBlock(final SpongeBlockSnapshot attachedSnapshot, final BlockState newState,
-            final SpongeBlockChangeFlag blockChange
-        ) {
-            super(attachedSnapshot.getBlockPos(), (BlockState) attachedSnapshot.getState());
-            this.original = attachedSnapshot;
-            this.newState = newState;
-            this.blockChangeFlag = blockChange;
-            this.originalOpacity = this.originalState.getOpacity(this.original.getServerWorld().get(), this.affectedPosition);
-        }
-
-        public BlockState getNewState() {
-            return this.newState;
-        }
-
-        public SpongeBlockChangeFlag getBlockChangeFlag() {
-            return this.blockChangeFlag;
-        }
-
-        public void populateChunkEffects(final ChunkPipeline.Builder builder) {
-
-            builder.addEffect(SetBlockToChunkSectionEffect.getInstance());
-            builder.addEffect(UpdateHeightMapEffect.getInstance());
-            builder.addEffect(UpdateChunkLightManagerEffect.getInstance());
-            builder.addEffect(OldBlockOnReplaceEffect.getInstance());
-            builder.addEffect(CheckBlockPostPlacementIsSameEffect.getInstance());
-            builder.addEffect(RefreshOldTileEntityOnChunkChangeEffect.getInstance());
-            builder.addEffect(BlockAddedEffect.getInstance());
-            builder.addEffect(UpdateOrCreateNewTileEntityPostPlacementEffect.getInstance());
-            builder.addEffect(ChunkChangeCompleteEffect.getInstance());
-        }
-
-        @Override
-        public Optional<BiConsumer<PhaseContext<@NonNull ?>, CauseStackManager.StackFrame>> getFrameMutator() {
-            return Optional.of((context, frame) -> {
-                context.addCreatorAndNotifierToCauseStack(frame);
-                frame.pushCause(this.original);
-            });
-        }
-
-        @Override
-        public boolean acceptTileAddition(final TileEntity tileEntity) {
-            if (!this.affectedPosition.equals(tileEntity.getPos())) {
-                return false;
-            }
-            if (this.queuedAdd != null) {
-                return false;
-
-            }
-            this.queuedAdd = tileEntity;
-            return true;
-        }
-
-        @Override
-        public boolean acceptTileRemoval(final TileEntity tileentity) {
-            if (!this.affectedPosition.equals(tileentity.getPos())) {
-                return false;
-            }
-            if (this.queuedRemoval != null) {
-                return false;
-            }
-            this.queuedRemoval = tileentity;
-            return true;
-        }
-
-        @Override
-        public boolean acceptTileReplacement(final @Nullable TileEntity existing, final TileEntity proposed) {
-            return existing != null && this.acceptTileRemoval(existing) && this.acceptTileAddition(proposed);
-        }
-
-        @Override
-        public void restore() {
-            this.original.restore(true, BlockChangeFlagManager.fromNativeInt(Constants.BlockChangeFlags.FORCED_RESTORE));
-        }
-
-        @Override
-        public void addToPrinter(final PrettyPrinter printer) {
-            printer.add("ChangeBlock")
-                .add(" %s : %s", "Original Block", this.original)
-                .add(" %s : %s", "New State", this.newState)
-                .add(" %s : %s", "RemovedTile", this.queuedRemoval)
-                .add(" %s : %s", "AddedTile", this.queuedAdd)
-                .add(" %s : %s", "ChangeFlag", this.blockChangeFlag);
-        }
-
-        @Override
-        protected SpongeBlockSnapshot getResultingSnapshot() {
-            return SpongeBlockSnapshotBuilder.pooled()
-                .world(this.original.getWorld())
-                .position(this.original.getPosition())
-                .blockState((org.spongepowered.api.block.BlockState) this.newState)
-                .build()
-                ;
-        }
-
-        @Override
-        protected SpongeBlockSnapshot getOriginalSnapshot() {
-            return this.original;
-        }
-    }
-
-
-    static final class NeighborNotification extends GameTransaction<NotifyNeighborBlockEvent> {
-        final BlockState original;
-        final BlockPos notifyPos;
-        final Block sourceBlock;
-        final BlockPos sourcePos;
-        private final Supplier<ServerWorld> serverWorld;
-        private Supplier<LocatableBlock> locatableBlock;
-
-        NeighborNotification(final Supplier<ServerWorld> serverWorldSupplier,
-            final BlockState notifyState, final BlockPos notifyPos,
-            final Block sourceBlock, final BlockPos sourcePos
-        ) {
-            super(sourcePos, notifyState);
-            this.serverWorld = serverWorldSupplier;
-            this.notifyPos = notifyPos;
-            this.sourceBlock = sourceBlock;
-            this.sourcePos = sourcePos;
-            this.original = serverWorldSupplier.get().getBlockState(sourcePos);
-            // This is one way to have lazily initialized fields
-            this.locatableBlock = () -> {
-                final LocatableBlock locatableBlock = new SpongeLocatableBlockBuilder()
-                    .world(this.serverWorld)
-                    .position(this.sourcePos.getX(), this.sourcePos.getY(), this.sourcePos.getZ())
-                    .state((org.spongepowered.api.block.BlockState) this.original)
-                    .build();
-                this.locatableBlock = () -> locatableBlock;
-                return locatableBlock;
-            };
-        }
-
-        @Override
-        public String toString() {
-            return new StringJoiner(", ", NeighborNotification.class.getSimpleName() + "[", "]")
-                .add("notifyState=" + this.originalState)
-                .add("notifyPos=" + this.notifyPos)
-                .add("sourceBlock=" + this.sourceBlock)
-                .add("sourcePos=" + this.sourcePos)
-                .add("actualSourceState=" + this.originalState)
-                .toString();
-        }
-
-        @Override
-        public Optional<BiConsumer<PhaseContext<@NonNull ?>, CauseStackManager.StackFrame>> getFrameMutator() {
-            return Optional.of((context, frame) -> {
-                frame.pushCause(this.locatableBlock.get());
-            });
-        }
-
-        @Override
-        public void addToPrinter(final PrettyPrinter printer) {
-            printer.add("NeighborNotification")
-                .add(" %s : %s, %s", "Source Pos", this.sourceBlock, this.sourcePos)
-                .add(" %s : %s, %s", "Notification", this.originalState, this.notifyPos);
-        }
-
-        @Override
-        public NotifyNeighborBlockEvent generateEvent(final PhaseContext<@NonNull ?> context,
-            final ImmutableList<GameTransaction<NotifyNeighborBlockEvent>> transactions,
-            final Cause currentCause
-        ) {
-            // TODO - for all neighbor notifications in the transactions find the direction of notification being used and pump into map.
-            final NotifyNeighborBlockEvent neighborBlockEvent = SpongeEventFactory.createNotifyNeighborBlockEvent(
-                currentCause,
-                new EnumMap<>(Direction.class),
-                Collections.emptyMap()
-            );
-            return neighborBlockEvent;
-        }
-
-        @Override
-        public void restore() {
-
-        }
-
-        @Override
-        public boolean canBatchWith(@Nullable final GameTransaction<@NonNull ?> next) {
-            return next instanceof NeighborNotification;
-        }
-
-        @Override
-        public boolean avoidsEvent() {
-            return true;
-        }
-
-        @Override
-        public boolean markCancelledTransactions(final NotifyNeighborBlockEvent event,
-            final ImmutableList<? extends GameTransaction<NotifyNeighborBlockEvent>> blockTransactions
-        ) {
-            return false;
-        }
-
-    }
-
-    static final class TransactionProcessState implements IPhaseState<TransactionContext> {
-
-        public static final TransactionProcessState TRANSACTION_PROCESS = new TransactionProcessState();
-
-        private TransactionProcessState() {
-        }
-
-        @Override
-        public TransactionContext createPhaseContext(final PhaseTracker server) {
-            throw new IllegalStateException("Cannot create context");
-        }
-
-        @Override
-        public void unwind(final TransactionContext phaseContext) {
-
-        }
-
-        @Override
-        public boolean tracksCreatorsAndNotifiers() {
-            return false;
-        }
-
-        @Override
-        public boolean isRestoring() {
-            return true;
-        }
-
-        @Override
-        public boolean doesBulkBlockCapture(final TransactionContext context) {
-            return false;
-        }
-
-        @Override
-        public boolean doesBlockEventTracking(final TransactionContext context) {
-            return false;
-        }
-
-        @Override
-        public boolean shouldCaptureBlockChangeOrSkip(
-            final TransactionContext phaseContext, final BlockPos pos, final BlockState currentState, final BlockState newState,
-            final BlockChangeFlag flags) {
-            return false;
-        }
-    }
-
-    static final class TransactionContext extends PhaseContext<TransactionContext> {
-
-        protected TransactionContext() {
-            super(TransactionProcessState.TRANSACTION_PROCESS, PhaseTracker.SERVER);
-        }
-    }
 }
