+++ src/main/java/org/spongepowered/common/event/tracking/context/transaction/pipeline/ChunkPipeline.java
@@ -33,8 +33,8 @@ import net.minecraft.world.server.ServerWorld;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.spongepowered.common.event.tracking.PhaseContext;
+import org.spongepowered.common.event.tracking.context.transaction.ChangeBlock;
 import org.spongepowered.common.event.tracking.context.transaction.EffectTransactor;
-import org.spongepowered.common.event.tracking.context.transaction.GameTransaction;
 import org.spongepowered.common.event.tracking.context.transaction.ResultingTransactionBySideEffect;
 import org.spongepowered.common.event.tracking.context.transaction.effect.EffectResult;
 import org.spongepowered.common.event.tracking.context.transaction.effect.PipelineCursor;
@@ -55,7 +55,7 @@ public final class ChunkPipeline implements BlockPipeline {
     private final @Nullable Supplier<ChunkSection> sectionSupplier;
     private final boolean wasEmpty;
     private final List<ResultingTransactionBySideEffect> chunkEffects;
-    final GameTransaction.ChangeBlock transaction;
+    final ChangeBlock transaction;
 
     ChunkPipeline(final Builder builder) {
         this.chunkSupplier = builder.chunkSupplier;
@@ -131,10 +131,10 @@ public final class ChunkPipeline implements BlockPipeline {
         @Nullable Supplier<Chunk> chunkSupplier;
         @Nullable Supplier<ChunkSection> sectionSupplier;
         boolean wasSectionEmpty;
-        GameTransaction.@MonotonicNonNull ChangeBlock transaction;
+        @MonotonicNonNull ChangeBlock transaction;
         List<ResultingTransactionBySideEffect> effects;
 
-        public Builder kickOff(final GameTransaction.ChangeBlock transaction) {
+        public Builder kickOff(final ChangeBlock transaction) {
             this.transaction = Objects.requireNonNull(transaction, "ChangeBlock transaction cannot be null!");
             return this;
         }
